add: Problem Solve BJ_G4_합이 0

+++ src/BJ_Practice/Silver/BJ_S1_1629_곱셈.java
@@ -18,7 +18,7 @@ public class BJ_S1_1629_곱셈 {
 
 		
 
-		System.out.println(power((long)A,B));
+		System.out.println(power(A,B));
 	}
 
 	static long power(long a, int b) {
